+++ includes/LknWCGatewayCieloCredit.php
@@ -18,7 +18,7 @@ use WC_Subscriptions_Order;
  */
 
 // Exit if accessed directly.
-if ( ! defined('ABSPATH')) {
+if (! defined('ABSPATH')) {
     exit;
 }
 
@@ -29,7 +29,8 @@ if ( ! defined('ABSPATH')) {
  *
  * @version  1.0.0
  */
-final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
+final class LknWCGatewayCieloCredit extends WC_Payment_Gateway
+{
     /**
      * Define instructions to configure and use this plugin.
      *
@@ -60,7 +61,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
     /**
      * Constructor for the gateway.
      */
-    public function __construct() {
+    public function __construct()
+    {
         $this->id = 'lkn_cielo_credit';
         $this->icon = apply_filters('lkn_wc_cielo_gateway_icon', '');
         $this->has_fields = true;
@@ -105,14 +107,16 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      * @param  WC_Order  $order
      * @return void
      */
-    public function process_subscription_payment($amount, $order): void {
+    public function process_subscription_payment($amount, $order): void
+    {
         do_action('lkn_wc_cielo_scheduled_subscription_payment', $amount, $order);
     }
 
     /**
      * Load admin JavaScript for the admin page.
      */
-    public function admin_load_script(): void {
+    public function admin_load_script(): void
+    {
         wp_enqueue_script('lkn-wc-gateway-admin', plugin_dir_url(__FILE__) . '../resources/js/admin/lkn-wc-gateway-admin.js', array('wp-i18n'), $this->version, 'all');
 
         $page = isset($_GET['page']) ? sanitize_text_field(wp_unslash($_GET['page'])) : '';
@@ -128,12 +132,27 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
                 'alertText' => __('Deseja realmente deletar todos logs dos pedidos?', 'lkn-wc-gateway-cielo')
             ));
         }
+
+        $allowed_sections = [
+            'lkn_cielo_credit',
+        ];
+
+        if (isset($_GET['section']) && in_array($_GET['section'], $allowed_sections, true)) {
+            wp_enqueue_script(
+                'lkn-wc-gateway-cielo-plugin-rate',
+                plugin_dir_url(__FILE__) . '../resources/js/admin/lkn-wc-gateway-plugin-rate.js',
+                array('jquery'),
+                $this->version,
+                false
+            );
+        }
     }
 
     /**
      * Load gateway scripts/styles.
      */
-    public function payment_gateway_scripts(): void {
+    public function payment_gateway_scripts(): void
+    {
         // Don't load scripts outside payment page
         if (
             ! is_checkout()
@@ -167,7 +186,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
     /**
      * Initialise Gateway Settings Form Fields.
      */
-    public function init_form_fields(): void {
+    public function init_form_fields(): void
+    {
         $this->form_fields = array(
             'general' => array(
                 'title' => esc_attr__('General', 'lkn-wc-gateway-cielo'),
@@ -302,7 +322,7 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
 
         $customConfigs = apply_filters('lkn_wc_cielo_get_custom_configs', array(), $this->id);
 
-        if ( ! empty($customConfigs)) {
+        if (! empty($customConfigs)) {
             $this->form_fields = array_merge($this->form_fields, $customConfigs);
         }
     }
@@ -310,10 +330,11 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
     /**
      * Render the payment fields.
      */
-    public function payment_fields(): void {
+    public function payment_fields(): void
+    {
         $activeInstallment = $this->get_option('installment_payment');
         $total_cart = number_format($this->get_order_total(), 2, '.', '');
-        $noLoginCheckout = isset($_GET['pay_for_order']) ? sanitize_text_field($_GET['pay_for_order']) : 'false';
+        $noLoginCheckout = isset($_GET['pay_for_order']) ? sanitize_text_field(wp_unslash($_GET['pay_for_order'])) : 'false';
         $installmentLimit = $this->get_option('installment_limit', 12);
         $installments = array();
         $nonce = wp_create_nonce('nonce_lkn_cielo_credit');
@@ -335,7 +356,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
 
         if ('yes' === $activeInstallment) {
             if (isset($_GET['pay_for_order'])) {
-                $order_id = wc_get_order_id_by_order_key(sanitize_text_field($_GET['key']));
+                $key = isset($_GET['key']) ? sanitize_text_field(wp_unslash($_GET['key'])) : '';
+                $order_id = wc_get_order_id_by_order_key($key);
                 $order = wc_get_order($order_id);
                 $total_cart = number_format($order->get_total(), 2, '.', '');
             }
@@ -476,17 +498,20 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    public function validate_fields() {
+    public function validate_fields()
+    {
         $validateCompatMode = $this->get_option('input_validation_compatibility', 'no');
-        if ( ! wp_verify_nonce($_POST['nonce_lkn_cielo_credit'], 'nonce_lkn_cielo_credit')) {
-            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($_POST['nonce_lkn_cielo_credit'], true), array('source' => 'woocommerce-cielo-credit'));
+        $nonce = isset($_POST['nonce_lkn_cielo_credit']) ? sanitize_text_field(wp_unslash($_POST['nonce_lkn_cielo_credit'])) : '';
+
+        if (! wp_verify_nonce($nonce, 'nonce_lkn_cielo_credit')) {
+            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($nonce, true), array('source' => 'woocommerce-cielo-credit'));
             $this->add_notice_once(__('Nonce verification failed, try reloading the page', 'lkn-wc-gateway-cielo'), 'error');
             return false;
         }
         if ('no' === $validateCompatMode) {
-            $ccnum = sanitize_text_field($_POST['lkn_ccno']);
-            $expDate = sanitize_text_field($_POST['lkn_cc_expdate']);
-            $cvv = sanitize_text_field($_POST['lkn_cc_cvc']);
+            $ccnum = isset($_POST['lkn_ccno']) ? sanitize_text_field(wp_unslash($_POST['lkn_ccno'])) : '';
+            $expDate = isset($_POST['lkn_cc_expdate']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_expdate'])) : '';
+            $cvv = isset($_POST['lkn_cc_cvc']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_cvc'])) : '';
 
             $validCcNumber = $this->validate_card_number($ccnum, true);
             $validExpDate = $this->validate_exp_date($expDate, true);
@@ -509,11 +534,13 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return array
      */
-    public function process_payment($order_id) {
+    public function process_payment($order_id)
+    {
         $nonceInactive = $this->get_option('nonce_compatibility', 'no');
+        $nonce = isset($_POST['nonce_lkn_cielo_credit']) ? sanitize_text_field(wp_unslash($_POST['nonce_lkn_cielo_credit'])) : '';
 
-        if ( ! wp_verify_nonce($_POST['nonce_lkn_cielo_credit'], 'nonce_lkn_cielo_credit') && 'no' === $nonceInactive) {
-            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($_POST['nonce_lkn_cielo_credit'], true), array('source' => 'woocommerce-cielo-credit'));
+        if (! wp_verify_nonce($nonce, 'nonce_lkn_cielo_credit') && 'no' === $nonceInactive) {
+            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($nonce, true), array('source' => 'woocommerce-cielo-credit'));
             $this->add_notice_once(__('Nonce verification failed, try reloading the page', 'lkn-wc-gateway-cielo'), 'error');
             throw new Exception(esc_attr(__('Nonce verification failed, try reloading the page', 'lkn-wc-gateway-cielo')));
         }
@@ -521,12 +548,12 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
         $order = wc_get_order($order_id);
 
         // Card parameters
-        $cardNum = preg_replace('/\s/', '', sanitize_text_field($_POST['lkn_ccno']));
-        $cardExpSplit = explode('/', preg_replace('/\s/', '', sanitize_text_field($_POST['lkn_cc_expdate'])));
+        $cardNum = preg_replace('/\s/', '', isset($_POST['lkn_ccno']) ? sanitize_text_field(wp_unslash($_POST['lkn_ccno'])) : '');
+        $cardExpSplit = explode('/', preg_replace('/\s/', '', isset($_POST['lkn_cc_expdate']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_expdate'])) : ''));
         $cardExp = $cardExpSplit[0] . '/20' . $cardExpSplit[1];
         $cardExpShort = $cardExpSplit[0] . '/' . $cardExpSplit[1];
-        $cardCvv = sanitize_text_field($_POST['lkn_cc_cvc']);
-        $cardName = sanitize_text_field($_POST['lkn_cc_cardholder_name']);
+        $cardCvv = isset($_POST['lkn_cc_cvc']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_cvc'])) : '';
+        $cardName = isset($_POST['lkn_cc_cardholder_name']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_cardholder_name'])) : '';
         $cardName = apply_filters('lkn_wc_cielo_get_cardholder_name', $cardName, $this, $order);
         $installments = 1;
 
@@ -592,7 +619,7 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
 
         // If installments option is active verify $_POST attribute
         if ('yes' === $activeInstallment) {
-            $installments = (int) sanitize_text_field($_POST['lkn_cc_installments']);
+            $installments = (int) isset($_POST['lkn_cc_installments']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_installments'])) : 1;
 
             if ($installments > 12) {
                 if (
@@ -693,7 +720,7 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
                 // Salvar o token e a bandeira do cartão no meta do pedido
                 $user_id = $order->get_user_id();
 
-                if ( ! isset($responseDecoded->Payment->CreditCard->CardToken)) {
+                if (! isset($responseDecoded->Payment->CreditCard->CardToken)) {
                     $order->add_order_note('O token para cobranças automáticas não foi gerado, então as cobranças automáticas não poderão ser efetuadas.');
                 }
 
@@ -773,7 +800,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
     /**
      * Calculate the total value of items in the WooCommerce cart.
      */
-    public static function lknGetCartTotal() {
+    public static function lknGetCartTotal()
+    {
         $cart = WC()->cart;
 
         if (empty($cart)) {
@@ -800,7 +828,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    public function process_refund($order_id, $amount = null, $reason = '') {
+    public function process_refund($order_id, $amount = null, $reason = '')
+    {
         // Do your refund here. Refund $amount for the order with ID $order_id
         $url = ($this->get_option('env') == 'production') ? 'https://api.cieloecommerce.cielo.com.br/' : 'https://apisandbox.cieloecommerce.cielo.com.br/';
         $merchantId = sanitize_text_field($this->get_option('merchant_id'));
@@ -844,7 +873,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    private function validate_card_number($ccnum, $renderNotice) {
+    private function validate_card_number($ccnum, $renderNotice)
+    {
         if (empty($ccnum)) {
             if ($renderNotice) {
                 $this->add_notice_once(__('Credit Card number is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -865,7 +895,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
         return true;
     }
 
-    private function validate_card_holder_name($cardName, $renderNotice) {
+    private function validate_card_holder_name($cardName, $renderNotice)
+    {
         if (empty($cardName) || strlen($cardName) < 3) {
             if ($renderNotice) {
                 $this->add_notice_once(__('Card Holder Name is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -885,7 +916,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    private function validate_exp_date($expDate, $renderNotice) {
+    private function validate_exp_date($expDate, $renderNotice)
+    {
         if (empty($expDate)) {
             if ($renderNotice) {
                 $this->add_notice_once(__('Expiration date is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -925,7 +957,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    private function validate_cvv($cvv, $renderNotice) {
+    private function validate_cvv($cvv, $renderNotice)
+    {
         if (empty($cvv)) {
             if ($renderNotice) {
                 $this->add_notice_once(__('CVV is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -952,8 +985,9 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      * @param string $message
      * @param string $type
      */
-    private function add_notice_once($message, $type): void {
-        if ( ! wc_has_notice($message, $type)) {
+    private function add_notice_once($message, $type): void
+    {
+        if (! wc_has_notice($message, $type)) {
             wc_add_notice($message, $type);
         }
     }
@@ -965,7 +999,8 @@ final class LknWCGatewayCieloCredit extends WC_Payment_Gateway {
      *
      * @return string|bool
      */
-    private function get_card_provider($cardNumber) {
+    private function get_card_provider($cardNumber)
+    {
         $brand = '';
         $brand = apply_filters('lkn_wc_cielo_get_card_brand', $brand, $cardNumber);