+++ includes/LknWCGatewayCieloDebit.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace Lkn\WCCieloPaymentGateway\Includes;
+
 use DateTime;
 use Exception;
 use Lkn\WCCieloPaymentGateway\Includes\LknWcCieloHelper;
@@ -16,7 +17,7 @@ use WC_Payment_Gateway;
  */
 
 // Exit if accessed directly.
-if ( ! defined('ABSPATH')) {
+if (! defined('ABSPATH')) {
     exit;
 }
 
@@ -27,7 +28,8 @@ if ( ! defined('ABSPATH')) {
  *
  * @version  1.0.0
  */
-final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
+final class LknWCGatewayCieloDebit extends WC_Payment_Gateway
+{
     /**
      * Define instructions to configure and use this plugin.
      *
@@ -59,7 +61,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
     /**
      * Constructor for the gateway.
      */
-    public function __construct() {
+    public function __construct()
+    {
         $this->id = 'lkn_cielo_debit';
         $this->icon = apply_filters('lkn_wc_cielo_gateway_icon', '');
         $this->has_fields = true;
@@ -82,7 +85,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         $this->instructions = $this->get_option('instructions', $this->description);
         $this->log = new WC_Logger();
         $gateway_enabled = get_option('woocommerce_' . $this->id . '_settings');
-        if($gateway_enabled['enabled'] === 'yes'){
+        if ($gateway_enabled['enabled'] === 'yes') {
             $this->accessToken = $this->generate_debit_auth_token();
         }
 
@@ -106,14 +109,15 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
     /**
      * Load admin JavaScript for the admin page.
      */
-    public function admin_load_script(): void {
+    public function admin_load_script(): void
+    {
         wp_enqueue_script('lkn-wc-gateway-admin', plugin_dir_url(__FILE__) . '../resources/js/admin/lkn-wc-gateway-admin.js', array('wp-i18n'), $this->version, 'all');
 
         $page = isset($_GET['page']) ? sanitize_text_field(wp_unslash($_GET['page'])) : '';
         $tab = isset($_GET['tab']) ? sanitize_text_field(wp_unslash($_GET['tab'])) : '';
         $section = isset($_GET['section']) ? sanitize_text_field(wp_unslash($_GET['section'])) : '';
 
-        if ( 'wc-settings' === $page && 'checkout' === $tab && $section == $this->id) {
+        if ('wc-settings' === $page && 'checkout' === $tab && $section == $this->id) {
             wp_enqueue_script('lknWCGatewayCieloDebitSettingsLayoutScript', plugin_dir_url(__FILE__) . '../resources/js/admin/lkn-wc-gateway-admin-layout.js', array('jquery'), $this->version, false);
             wp_enqueue_style('lkn-admin-layout', plugin_dir_url(__FILE__) . '../resources/css/frontend/lkn-admin-layout.css', array(), $this->version, 'all');
             wp_enqueue_script('lknWCGatewayCieloDebitClearButtonScript', plugin_dir_url(__FILE__) . '../resources/js/admin/lkn-clear-logs-button.js', array('jquery', 'wp-api'), $this->version, false);
@@ -122,12 +126,27 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
                 'alertText' => __('Deseja realmente deletar todos logs dos pedidos?', 'lkn-wc-gateway-cielo')
             ));
         }
+
+        $allowed_sections = [
+            'lkn_cielo_debit'
+        ];
+
+        if (isset($_GET['section']) && in_array($_GET['section'], $allowed_sections, true)) {
+            wp_enqueue_script(
+                'lkn-wc-gateway-cielo-plugin-rate',
+                plugin_dir_url(__FILE__) . '../resources/js/admin/lkn-wc-gateway-plugin-rate.js',
+                array('jquery'),
+                $this->version,
+                false
+            );
+        }
     }
 
     /**
      * Load gateway scripts/styles.
      */
-    public function payment_gateway_scripts(): void {
+    public function payment_gateway_scripts(): void
+    {
         // Don't load scripts outside payment page
         if (
             ! is_checkout()
@@ -166,17 +185,18 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         wp_enqueue_style('lkn-dc-style', plugin_dir_url(__FILE__) . '../resources/css/frontend/lkn-dc-style.css', array(), $this->version, 'all');
 
         wp_enqueue_style('lkn-mask', plugin_dir_url(__FILE__) . '../resources/css/frontend/lkn-mask.css', array(), $this->version, 'all');
-        
+
         wp_enqueue_script('lkn-fix-token-script', plugin_dir_url(__FILE__) . '../resources/js/frontend/lkn-fix-token-script.js', array('jquery', 'wp-api'), $this->version, false);
     }
 
     /**
      * Initialise Gateway Settings Form Fields.
      */
-    public function init_form_fields(): void {
+    public function init_form_fields(): void
+    {
         $this->form_fields = array(
             'general' => array(
-                'title' => esc_attr__( 'General', 'lkn-wc-gateway-cielo' ),
+                'title' => esc_attr__('General', 'lkn-wc-gateway-cielo'),
                 'type' => 'title',
             ),
             'enabled' => array(
@@ -292,7 +312,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
                 'desc_tip' => true,
             ),
             'card' => array(
-                'title' => esc_attr__( 'Card', 'lkn-wc-gateway-cielo' ),
+                'title' => esc_attr__('Card', 'lkn-wc-gateway-cielo'),
                 'type' => 'title',
             ),
             'installment_payment' => array(
@@ -330,7 +350,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
                 'default' => 'yes',
             ),
             'developer' => array(
-                'title' => esc_attr__( 'Developer', 'lkn-wc-gateway-cielo' ),
+                'title' => esc_attr__('Developer', 'lkn-wc-gateway-cielo'),
                 'type' => 'title',
             ),
             'debug' => array(
@@ -360,7 +380,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
 
         $customConfigs = apply_filters('lkn_wc_cielo_get_custom_configs', array(), $this->id);
 
-        if ( ! empty($customConfigs)) {
+        if (! empty($customConfigs)) {
             $this->form_fields = array_merge($this->form_fields, $customConfigs);
         }
     }
@@ -368,7 +388,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
     /**
      * Generate Cielo API 3.0 in auth token.
      */
-    public function generate_debit_auth_token() {
+    public function generate_debit_auth_token()
+    {
         try {
             $env = $this->get_option('env');
             $clientId = $this->get_option('client_id');
@@ -412,8 +433,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
                     'expires_in' => $responseDecoded->expires_in,
                 );
             }
-        } catch ( Exception $e ) {
-            $this->add_error( $e->getMessage() );
+        } catch (Exception $e) {
+            $this->add_error($e->getMessage());
             $debug = $this->get_option('debug');
 
             if ('yes' === $debug) {
@@ -426,7 +447,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
     /**
      * Calculate the total value of items in the WooCommerce cart.
      */
-    public static function lknGetCartTotal() {
+    public static function lknGetCartTotal()
+    {
         $cart = WC()->cart;
 
         if (empty($cart)) {
@@ -444,19 +466,25 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         return 0;
     }
 
-    private function get_client_ip() {
+    private function get_client_ip()
+    {
         $ip_address = '';
+        $client_ip = isset($_SERVER['HTTP_CLIENT_IP']) ? sanitize_text_field(wp_unslash($_SERVER['HTTP_CLIENT_IP'])) : '';
+        $forwarded_ip = isset($_SERVER['HTTP_X_FORWARDED_FOR']) ? sanitize_text_field(wp_unslash($_SERVER['HTTP_X_FORWARDED_FOR'])) : '';
+        $real_ip = isset($_SERVER['HTTP_X_REAL_IP']) ? sanitize_text_field(wp_unslash($_SERVER['HTTP_X_REAL_IP'])) : '';
+        $remote_ip = isset($_SERVER['REMOTE_ADDR']) ? sanitize_text_field(wp_unslash($_SERVER['REMOTE_ADDR'])) : '';
 
-        if ( ! empty($_SERVER['HTTP_CLIENT_IP'])) {
-            $ip_address = $_SERVER['HTTP_CLIENT_IP'];
-        } elseif ( ! empty($_SERVER['HTTP_X_FORWARDED_FOR'])) {
+
+        if (! empty($client_ip)) {
+            $ip_address = $client_ip;
+        } elseif (! empty($forwarded_ip)) {
             // Se estiver atrás de um proxy, `HTTP_X_FORWARDED_FOR` pode conter uma lista de IPs.
-            $ip_list = explode(',', $_SERVER['HTTP_X_FORWARDED_FOR']);
+            $ip_list = explode(',', $forwarded_ip);
             $ip_address = trim($ip_list[0]); // Pega o primeiro IP da lista
-        } elseif ( ! empty($_SERVER['HTTP_X_REAL_IP'])) {
-            $ip_address = $_SERVER['HTTP_X_REAL_IP'];
+        } elseif (! empty($real_ip)) {
+            $ip_address = $real_ip;
         } else {
-            $ip_address = $_SERVER['REMOTE_ADDR'];
+            $ip_address = $remote_ip;
         }
 
         return $ip_address;
@@ -465,11 +493,12 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
     /**
      * Render the payment fields.
      */
-    public function payment_fields(): void {
+    public function payment_fields(): void
+    {
         $total_cart = number_format($this->get_order_total(), 2, '', '');
         $accessToken = $this->accessToken;
         $url = get_page_link();
-        $nonce = wp_create_nonce( 'nonce_lkn_cielo_debit');
+        $nonce = wp_create_nonce('nonce_lkn_cielo_debit');
         $placeholder = $this->get_option('placeholder', 'no');
         $placeholderEnabled = false;
 
@@ -478,7 +507,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         }
 
         $activeInstallment = $this->get_option('installment_payment');
-        $noLoginCheckout = isset($_GET['pay_for_order']) ? sanitize_text_field($_GET['pay_for_order']) : 'false';
+        $noLoginCheckout = isset($_GET['pay_for_order']) ? sanitize_text_field(wp_unslash($_GET['pay_for_order'])) : 'false';
         $installmentLimit = $this->get_option('installment_limit', 12);
         $installments = array();
         $installmentsTotal = number_format($this->get_order_total(), 2, '.', '');
@@ -486,7 +515,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         $installmentLimit = apply_filters('lkn_wc_cielo_set_installment_limit', $installmentLimit, $this);
 
         if (isset($_GET['pay_for_order'])) {
-            $order_id = wc_get_order_id_by_order_key(sanitize_text_field($_GET['key']));
+            $key = isset($_GET['key']) ? sanitize_text_field(wp_unslash($_GET['key'])) : '';
+            $order_id = wc_get_order_id_by_order_key($key);
             $order = wc_get_order($order_id);
             $total_cart = number_format($order->get_total(), 2, '', '');
         }
@@ -961,17 +991,20 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    public function validate_fields() {
+    public function validate_fields()
+    {
         $validateCompatMode = $this->get_option('input_validation_compatibility', 'no');
-        if ( ! wp_verify_nonce($_POST['nonce_lkn_cielo_debit'], 'nonce_lkn_cielo_debit')) {
-            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($_POST['nonce_lkn_cielo_debit'], true), array('source' => 'woocommerce-cielo-debit'));
+        $nonce = isset($_POST['nonce_lkn_cielo_debit']) ? sanitize_text_field(wp_unslash($_POST['nonce_lkn_cielo_debit'])) : '';
+
+        if (! wp_verify_nonce($nonce, 'nonce_lkn_cielo_debit')) {
+            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($nonce, true), array('source' => 'woocommerce-cielo-debit'));
             $this->add_notice_once(__('Nonce verification failed, try reloading the page', 'lkn-wc-gateway-cielo'), 'error');
             return false;
         }
         if ('no' === $validateCompatMode) {
-            $dcnum = sanitize_text_field($_POST['lkn_dcno']);
-            $expDate = sanitize_text_field($_POST['lkn_dc_expdate']);
-            $cvv = sanitize_text_field($_POST['lkn_dc_cvc']);
+            $dcnum = isset($_POST['lkn_dcno']) ? sanitize_text_field(wp_unslash($_POST['lkn_dcno'])) : '';
+            $expDate = isset($_POST['lkn_dc_expdate']) ? sanitize_text_field(wp_unslash($_POST['lkn_dc_expdate'])) : '';
+            $cvv = isset($_POST['lkn_dc_cvc']) ? sanitize_text_field(wp_unslash($_POST['lkn_dc_cvc'])) : '';
 
             $validdcNumber = $this->validate_card_number($dcnum, true);
             $validExpDate = $this->validate_exp_date($expDate, true);
@@ -994,10 +1027,13 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return array
      */
-    public function process_payment($order_id) {
+    public function process_payment($order_id)
+    {
         $nonceInactive = $this->get_option('nonce_compatibility', 'no');
-        if ( ! wp_verify_nonce($_POST['nonce_lkn_cielo_debit'], 'nonce_lkn_cielo_debit') && 'no' === $nonceInactive) {
-            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($_POST['nonce_lkn_cielo_debit'], true), array('source' => 'woocommerce-cielo-debit'));
+        $nonce = isset($_POST['nonce_lkn_cielo_debit']) ? sanitize_text_field(wp_unslash($_POST['nonce_lkn_cielo_debit'])) : '';
+
+        if (! wp_verify_nonce($nonce, 'nonce_lkn_cielo_debit') && 'no' === $nonceInactive) {
+            $this->log->log('error', 'Nonce verification failed. Nonce: ' . var_export($nonce, true), array('source' => 'woocommerce-cielo-debit'));
             $this->add_notice_once(__('Nonce verification failed, try reloading the page', 'lkn-wc-gateway-cielo'), 'error');
             throw new Exception(esc_attr(__('Nonce verification failed, try reloading the page', 'lkn-wc-gateway-cielo')));
         }
@@ -1005,22 +1041,22 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         $order = wc_get_order($order_id);
 
         // Card parameters
-        $cardNum = preg_replace('/\s/', '', sanitize_text_field($_POST['lkn_dcno']));
-        $cardExpSplit = explode('/', preg_replace('/\s/', '', sanitize_text_field($_POST['lkn_dc_expdate'])));
+        $cardNum = preg_replace('/\s/', '', isset($_POST['lkn_dcno']) ? sanitize_text_field(wp_unslash($_POST['lkn_dcno'])) : '');
+        $cardExpSplit = explode('/', preg_replace('/\s/', '', isset($_POST['lkn_dc_expdate']) ? sanitize_text_field(wp_unslash($_POST['lkn_dc_expdate'])) : ''));
         $cardExp = $cardExpSplit[0] . '/20' . $cardExpSplit[1];
         $cardExpShort = $cardExpSplit[0] . '/' . $cardExpSplit[1];
-        $cardCvv = sanitize_text_field($_POST['lkn_dc_cvc']);
-        $cardName = sanitize_text_field($_POST['lkn_dc_cardholder_name']);
+        $cardCvv = isset($_POST['lkn_dc_cvc']) ? sanitize_text_field(wp_unslash($_POST['lkn_dc_cvc'])) : '';
+        $cardName = isset($_POST['lkn_dc_cardholder_name']) ? sanitize_text_field(wp_unslash($_POST['lkn_dc_cardholder_name'])) : '';
         $cardName = apply_filters('lkn_wc_cielo_get_cardholder_name', $cardName, $this, $order);
-        $cardType = sanitize_text_field($_POST['lkn_cc_type']);
+        $cardType = isset($_POST['lkn_cc_type']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_type'])) : '';
         $installments = 1;
-
+        
         // Authentication parameters
-        $xid = sanitize_text_field($_POST['lkn_cielo_3ds_xid']);
-        $cavv = sanitize_text_field($_POST['lkn_cielo_3ds_cavv']);
-        $eci = sanitize_text_field($_POST['lkn_cielo_3ds_eci']);
-        $version = sanitize_text_field($_POST['lkn_cielo_3ds_version']);
-        $refId = sanitize_text_field($_POST['lkn_cielo_3ds_ref_id']);
+        $xid = isset($_POST['lkn_cielo_3ds_xid']) ? sanitize_text_field(wp_unslash($_POST['lkn_cielo_3ds_xid'])) : '';
+        $cavv = isset($_POST['lkn_cielo_3ds_cavv']) ? sanitize_text_field(wp_unslash($_POST['lkn_cielo_3ds_cavv'])) : '';
+        $eci = isset($_POST['lkn_cielo_3ds_eci']) ? sanitize_text_field(wp_unslash($_POST['lkn_cielo_3ds_eci'])) : '';
+        $version = isset($_POST['lkn_cielo_3ds_version']) ? sanitize_text_field(wp_unslash($_POST['lkn_cielo_3ds_version'])) : '';
+        $refId = isset($_POST['lkn_cielo_3ds_ref_id']) ? sanitize_text_field(wp_unslash($_POST['lkn_cielo_3ds_ref_id'])) : '';
 
         // POST parameters
         $url = ($this->get_option('env') == 'production') ? 'https://api.cieloecommerce.cielo.com.br/' : 'https://apisandbox.cieloecommerce.cielo.com.br/';
@@ -1080,7 +1116,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
 
         // If installments option is active verify $_POST attribute
         if ('yes' === $activeInstallment && 'Credit' == $cardType) {
-            $installments = (int) sanitize_text_field($_POST['lkn_cc_installments']);
+            $installments = (int) isset($_POST['lkn_cc_installments']) ? sanitize_text_field(wp_unslash($_POST['lkn_cc_installments'])) : 1;
 
             if ($installments > 12) {
                 if (
@@ -1234,7 +1270,7 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         }
 
         $response = wp_remote_post($url . '1/sales', $args);
-       /*  throw new Exception(json_encode($args)); */
+        /*  throw new Exception(json_encode($args)); */
         if (is_wp_error($response)) {
             if ('yes' === $debug) {
                 $this->log->log('error', var_export($response->get_error_messages(), true), array('source' => 'woocommerce-cielo-debit'));
@@ -1328,7 +1364,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
         throw new Exception(esc_attr($message));
     }
 
-    private function validate_card_holder_name($cardName, $renderNotice) {
+    private function validate_card_holder_name($cardName, $renderNotice)
+    {
         if (empty($cardName) || strlen($cardName) < 3) {
             if ($renderNotice) {
                 $this->add_notice_once(__('Card Holder Name is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -1349,7 +1386,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    public function process_refund($order_id, $amount = null, $reason = '') {
+    public function process_refund($order_id, $amount = null, $reason = '')
+    {
         // Do your refund here. Refund $amount for the order with ID $order_id
         $url = ($this->get_option('env') == 'production') ? 'https://api.cieloecommerce.cielo.com.br/' : 'https://apisandbox.cieloecommerce.cielo.com.br/';
         $merchantId = sanitize_text_field($this->get_option('merchant_id'));
@@ -1393,7 +1431,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    private function validate_card_number($dcnum, $renderNotice) {
+    private function validate_card_number($dcnum, $renderNotice)
+    {
         if (empty($dcnum)) {
             if ($renderNotice) {
                 $this->add_notice_once(__('Debit Card number is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -1422,7 +1461,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    private function validate_exp_date($expDate, $renderNotice) {
+    private function validate_exp_date($expDate, $renderNotice)
+    {
         if (empty($expDate)) {
             if ($renderNotice) {
                 $this->add_notice_once(__('Expiration date is required!', 'lkn-wc-gateway-cielo'), 'error');
@@ -1460,7 +1500,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return bool
      */
-    private function validate_cvv($cvv, $renderNotice) {
+    private function validate_cvv($cvv, $renderNotice)
+    {
         if (empty($cvv)) {
             $this->add_notice_once(__('CVV is required!', 'lkn-wc-gateway-cielo'), 'error');
 
@@ -1483,8 +1524,9 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      * @param string $message
      * @param string $type
      */
-    private function add_notice_once($message, $type): void {
-        if ( ! wc_has_notice($message, $type)) {
+    private function add_notice_once($message, $type): void
+    {
+        if (! wc_has_notice($message, $type)) {
             wc_add_notice($message, $type);
         }
     }
@@ -1496,7 +1538,8 @@ final class LknWCGatewayCieloDebit extends WC_Payment_Gateway {
      *
      * @return string|bool
      */
-    private function get_card_provider($cardNumber) {
+    private function get_card_provider($cardNumber)
+    {
         $brand = '';
         $brand = apply_filters('lkn_wc_cielo_get_card_brand', $brand, $cardNumber);