+++ lkn-wc-gateway-cielo.php
@@ -3,7 +3,7 @@
  * Plugin Name: Payment Gateway for Cielo API on WooCommerce
  * Plugin URI: https://www.linknacional.com.br/wordpress/woocommerce/cielo/
  * Description: Adds the Cielo API 3.0 Payments gateway to your WooCommerce website.
- * Version: 1.15.0
+ * Version: 1.16.0
  * Author: Link Nacional
  * Author URI: https://linknacional.com.br
  * Text Domain: lkn-wc-gateway-cielo
@@ -22,7 +22,7 @@ use Lkn\WCCieloPaymentGateway\Includes\LknWcCieloCreditBlocks;
 use Lkn\WCCieloPaymentGateway\Includes\LknWcCieloDebitBlocks;
 
 // Exit if accessed directly.
-if ( ! defined('ABSPATH')) {
+if (! defined('ABSPATH')) {
     exit;
 }
 
@@ -33,13 +33,15 @@ require_once __DIR__ . '/vendor/autoload.php';
  *
  * @class Lkn_WC_Cielo_Payment
  */
-final class LknWCCieloPayment {
+final class LknWCCieloPayment
+{
     /**
      * Show plugin dependency notice.
      *
      * @since
      */
-    public static function __lkn_wc_gateway_cielo_dependency_notice(): void {
+    public static function __lkn_wc_gateway_cielo_dependency_notice(): void
+    {
         // Admin notice.
         $message = sprintf(
             '<strong>%1$s</strong> %2$s <a href="%3$s" target="_blank">%4$s</a>  %5$s %6$s+ %7$s.',
@@ -60,7 +62,8 @@ final class LknWCCieloPayment {
      *
      * @since
      */
-    public static function __lkn_wc_gateway_cielo_inactive_notice(): void {
+    public static function __lkn_wc_gateway_cielo_inactive_notice(): void
+    {
         // Admin notice.
         $message = sprintf(
             '<div class="notice notice-error"><p><strong>%1$s</strong> %2$s <a href="%3$s" target="_blank">%4$s</a> %5$s.</p></div>',
@@ -77,10 +80,8 @@ final class LknWCCieloPayment {
     /**
      * Plugin bootstrapping.
      */
-    public static function init(): void {
-        // Load text domains
-        add_action('init', array(__CLASS__, 'lkn_wc_gateway_cielo_load_textdomain'));
-
+    public static function init(): void
+    {
         add_action('beforewoocommerce_init', array(__CLASS__, 'wcEditorBlocksActive'));
         add_action('woocommerce_blocks_payment_method_type_registration', array(__CLASS__, 'wcEditorBlocksAddPaymentMethod'));
 
@@ -105,10 +106,20 @@ final class LknWCCieloPayment {
         add_action('rest_api_init', array(new LknWCGatewayCieloEndpoint(), 'registerOrderCaptureEndPoint'));
 
         add_action('add_meta_boxes', array(new LknWcCieloHelper(), 'showOrderLogs'));
+
+        add_action('admin_notices', array(__CLASS__, 'lkn_admin_notice'));
     }
 
-    public static function wcEditorBlocksActive(): void {
-        if ( class_exists( '\Automattic\WooCommerce\Utilities\FeaturesUtil' ) ) {
+    public static function lkn_admin_notice()
+    {
+        if (!file_exists(WP_PLUGIN_DIR . '/fraud-detection-for-woocommerce/fraud-detection-for-woocommerce.php') && !is_plugin_active('integration-rede-for-woocommerce/integration-rede-for-woocommerce.php')) {
+            require LKN_WC_GATEWAY_CIELO_DIR . 'includes/views/notices/LknWcCieloDownloadNotice.php';
+        }
+    }
+
+    public static function wcEditorBlocksActive(): void
+    {
+        if (class_exists('\Automattic\WooCommerce\Utilities\FeaturesUtil')) {
             Automattic\WooCommerce\Utilities\FeaturesUtil::declare_compatibility(
                 'cart_checkout_blocks',
                 __FILE__,
@@ -117,13 +128,14 @@ final class LknWCCieloPayment {
         }
     }
 
-    public static function wcEditorBlocksAddPaymentMethod(Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry $payment_method_registry): void {
-        if ( ! class_exists( 'Automattic\WooCommerce\Blocks\Payments\Integrations\AbstractPaymentMethodType' )) {
+    public static function wcEditorBlocksAddPaymentMethod(Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry $payment_method_registry): void
+    {
+        if (! class_exists('Automattic\WooCommerce\Blocks\Payments\Integrations\AbstractPaymentMethodType')) {
             return;
         }
 
-        $payment_method_registry->register( new LknWcCieloCreditBlocks() );
-        $payment_method_registry->register( new LknWcCieloDebitBlocks() );
+        $payment_method_registry->register(new LknWcCieloCreditBlocks());
+        $payment_method_registry->register(new LknWcCieloDebitBlocks());
     }
 
     /**
@@ -133,14 +145,15 @@ final class LknWCCieloPayment {
      *
      * @since
      */
-    public static function check_environment() {
+    public static function check_environment()
+    {
         // Is not admin
-        if ( ! is_admin() || ! current_user_can('activate_plugins')) {
+        if (! is_admin() || ! current_user_can('activate_plugins')) {
             return null;
         }
 
         // Load plugin helper functions.
-        if ( ! function_exists('deactivate_plugins') || ! function_exists('is_plugin_active')) {
+        if (! function_exists('deactivate_plugins') || ! function_exists('is_plugin_active')) {
             require_once ABSPATH . '/wp-admin/includes/plugin.php';
         }
 
@@ -159,14 +172,14 @@ final class LknWCCieloPayment {
                     $is_installed = ! empty($all_plugins['woocommerce/woocommerce.php']);
                 }
 
-                if ( ! $is_installed) {
+                if (! $is_installed) {
                     add_action('admin_notices', array(__CLASS__, '__lkn_wc_gateway_cielo_dependency_notice'));
                 }
 
                 // Check for if give plugin activate or not.
                 $is_wc_active = class_exists('WooCommerce');
 
-                if ( ! $is_wc_active) {
+                if (! $is_wc_active) {
                     add_action('admin_notices', array(__CLASS__, '__lkn_wc_gateway_cielo_inactive_notice'));
 
                     $is_deactivate_plugin = true;
@@ -191,19 +204,13 @@ final class LknWCCieloPayment {
     }
 
     /**
-     * Load the plugin text domain.
-     */
-    public static function lkn_wc_gateway_cielo_load_textdomain(): void {
-        load_plugin_textdomain('lkn-wc-gateway-cielo', false, dirname(plugin_basename(__FILE__)) . '/languages');
-    }
-
-    /**
      * Add the Cielo Payment gateway to the list of available gateways.
      *
      * @param array
      * @param mixed $gateways
      */
-    public static function add_gateway($gateways) {
+    public static function add_gateway($gateways)
+    {
         $gateways[] = new LknWCGatewayCieloCredit();
         $gateways[] = new LknWCGatewayCieloDebit();
 
@@ -213,7 +220,8 @@ final class LknWCCieloPayment {
     /**
      * Plugin includes.
      */
-    public static function includes(): void {
+    public static function includes(): void
+    {
         LknWCCieloPayment::setup_constants();
         LknWCCieloPayment::check_environment();
     }
@@ -223,7 +231,8 @@ final class LknWCCieloPayment {
      *
      * @return string
      */
-    public static function plugin_url() {
+    public static function plugin_url()
+    {
         return untrailingslashit(plugins_url('/', __FILE__));
     }
 
@@ -232,7 +241,8 @@ final class LknWCCieloPayment {
      *
      * @return string
      */
-    public static function plugin_abspath() {
+    public static function plugin_abspath()
+    {
         return trailingslashit(plugin_dir_path(__FILE__));
     }
 
@@ -246,7 +256,8 @@ final class LknWCCieloPayment {
      *
      * @return array
      */
-    public static function lkn_wc_cielo_plugin_row_meta($plugin_meta, $plugin_file) {
+    public static function lkn_wc_cielo_plugin_row_meta($plugin_meta, $plugin_file)
+    {
         $new_meta_links['setting'] = sprintf(
             '<a href="%1$s">%2$s</a>',
             admin_url('admin.php?page=wc-settings&tab=checkout'),
@@ -256,7 +267,8 @@ final class LknWCCieloPayment {
         return array_merge($plugin_meta, $new_meta_links);
     }
 
-    public static function lkn_wc_cielo_plugin_row_meta_pro($plugin_meta, $plugin_file) {
+    public static function lkn_wc_cielo_plugin_row_meta_pro($plugin_meta, $plugin_file)
+    {
         // Defina o URL e o texto do link
         $url = 'https://www.linknacional.com.br/wordpress/woocommerce/cielo/';
         $link_text = sprintf(
@@ -278,7 +290,8 @@ final class LknWCCieloPayment {
      *
      * @param WC_Order $order
      */
-    public static function order_details_after_order_table($order): void {
+    public static function order_details_after_order_table($order): void
+    {
         $installment = $order->get_meta('installments');
 
         if ($installment && $installment > 1) {
@@ -293,7 +306,8 @@ final class LknWCCieloPayment {
      * @param bool     $sent_to_admin
      * @param WC_Order $order
      */
-    public static function email_order_meta_fields($fields, $sent_to_admin, $order) {
+    public static function email_order_meta_fields($fields, $sent_to_admin, $order)
+    {
         $installment = $order->get_meta('installments');
         if ($installment && $installment > 1) {
             $fields['installment'] = array(
@@ -308,24 +322,25 @@ final class LknWCCieloPayment {
     /**
      * Setup plugin constants for ease of use.
      */
-    private static function setup_constants(): void {
+    private static function setup_constants(): void
+    {
         // Defines addon version number for easy reference.
-        if ( ! defined('LKN_WC_CIELO_VERSION')) {
-            define('LKN_WC_CIELO_VERSION', '1.15.0');
+        if (! defined('LKN_WC_CIELO_VERSION')) {
+            define('LKN_WC_CIELO_VERSION', '1.16.0');
         }
-        if ( ! defined('LKN_WC_CIELO_TRANSLATION_PATH')) {
+        if (! defined('LKN_WC_CIELO_TRANSLATION_PATH')) {
             define('LKN_WC_CIELO_TRANSLATION_PATH', plugin_dir_path(__FILE__) . 'languages/');
         }
-        if ( ! defined('LKN_WC_GATEWAY_CIELO_BASENAME')) {
+        if (! defined('LKN_WC_GATEWAY_CIELO_BASENAME')) {
             define('LKN_WC_GATEWAY_CIELO_BASENAME', plugin_basename(__FILE__));
         }
-        if ( ! defined('LKN_WC_GATEWAY_CIELO_DIR')) {
+        if (! defined('LKN_WC_GATEWAY_CIELO_DIR')) {
             define('LKN_WC_GATEWAY_CIELO_DIR', plugin_dir_path(__FILE__));
         }
-        if ( ! defined('LKN_WC_GATEWAY_CIELO_URL')) {
+        if (! defined('LKN_WC_GATEWAY_CIELO_URL')) {
             define('LKN_WC_GATEWAY_CIELO_URL', plugin_dir_url(__FILE__));
         }
-        if ( ! defined('LKN_WC_GATEWAY_CIELO_MIN_WC_VERSION')) {
+        if (! defined('LKN_WC_GATEWAY_CIELO_MIN_WC_VERSION')) {
             define('LKN_WC_GATEWAY_CIELO_MIN_WC_VERSION', '5.0.0');
         }
     }